+++ mitra_activitypub/src/actors/handlers.rs
@@ -85,7 +85,10 @@ use crate::{
         SERVICE,
         VERIFIABLE_IDENTITY_STATEMENT,
     },
-    webfinger::perform_webfinger_query,
+    webfinger::{
+        perform_webfinger_query,
+        peform_reverse_webfinger_query,
+    },
 };
 
 use super::{
@@ -272,15 +275,28 @@ async fn get_webfinger_hostname(
     local_hostname: &str,
     actor: &ValidatedActor,
     has_portable_account: bool,
-) -> Result<WebfingerHostname, HandlerError> {
+) -> Result<(Option<String>, WebfingerHostname), HandlerError> {
     let canonical_actor_id = CanonicalUri::parse(&actor.id)
         .map_err(|_| ValidationError("invalid actor ID"))?;
-    let webfinger_hostname = match canonical_actor_id {
+    let (server_hostname, webfinger_hostname) = match canonical_actor_id {
         CanonicalUri::Http(http_uri) => {
-            // TODO: implement reverse webfinger lookup
-            // https://swicg.github.io/activitypub-webfinger/#reverse-discovery
-            let hostname = http_uri.hostname().to_string();
-            WebfingerHostname::Remote(hostname)
+            let server_hostname = http_uri.hostname().to_string();
+            let webfinger_hostname = match peform_reverse_webfinger_query(
+                agent,
+                &actor.preferred_username,
+                &http_uri,
+            ).await {
+                Ok(webfinger_hostname) => webfinger_hostname,
+                Err(error) => {
+                    log::warn!(
+                        "reverse webfinger query failed for {}: {}",
+                        http_uri,
+                        error,
+                    );
+                    server_hostname.clone()
+                },
+            };
+            (Some(server_hostname), WebfingerHostname::Remote(webfinger_hostname))
         },
         CanonicalUri::Ap(_) => {
             if let Some(gateway) = actor.gateways.first() {
@@ -290,10 +306,10 @@ async fn get_webfinger_hostname(
                 if gateway_hostname == local_hostname {
                     // Portable actor with local account (unmanaged)
                     if has_portable_account {
-                        return Ok(WebfingerHostname::Local);
+                        return Ok((None, WebfingerHostname::Local));
                     } else {
                         // WARNING: only allowed when profile is being created
-                        return Ok(WebfingerHostname::Unknown);
+                        return Ok((None, WebfingerHostname::Unknown));
                     };
                 };
                 let webfinger_address = WebfingerAddress::new_unchecked(
@@ -308,7 +324,7 @@ async fn get_webfinger_hostname(
                     .map_err(|_| ValidationError("invalid actor ID in JRD"))?;
                 if canonical_webfinger_actor_id == canonical_actor_id {
                     // Actor is hosted by this gateway
-                    WebfingerHostname::Remote(gateway_hostname)
+                    (Some(gateway_hostname.clone()), WebfingerHostname::Remote(gateway_hostname))
                 } else {
                     return Err(ValidationError("unexpected actor ID in JRD").into());
                 }
@@ -317,7 +333,7 @@ async fn get_webfinger_hostname(
             }
         },
     };
-    Ok(webfinger_hostname)
+    Ok((server_hostname, webfinger_hostname))
 }
 
 enum ActorImageResult {
@@ -641,7 +657,7 @@ pub async fn create_remote_profile(
     actor: Actor,
 ) -> Result<DbActorProfile, HandlerError> {
     let Actor { inner: actor, value: actor_json } = actor;
-    let webfinger_hostname = get_webfinger_hostname(
+    let (server_hostname, webfinger_hostname) = get_webfinger_hostname(
         &ap_client.agent(),
         &ap_client.instance.hostname(),
         &actor,
@@ -680,7 +696,8 @@ pub async fn create_remote_profile(
     ).await?;
     let mut profile_data = ProfileCreateData {
         username: actor.preferred_username.clone(),
-        hostname: webfinger_hostname,
+        hostname: server_hostname,
+        webfinger_hostname: webfinger_hostname,
         display_name: actor.name.clone(),
         bio: actor.summary.clone(),
         avatar: maybe_avatar.ok(),
@@ -723,7 +740,7 @@ pub async fn update_remote_profile(
     let actor_data_old = profile.expect_actor_data();
     let actor_data = actor.to_db_actor()?;
     assert_eq!(actor_data_old.id, actor_data.id, "actor ID shouldn't change");
-    let webfinger_hostname = get_webfinger_hostname(
+    let (server_hostname, webfinger_hostname) = get_webfinger_hostname(
         &ap_client.agent(),
         &ap_client.instance.hostname(),
         &actor,
@@ -754,7 +771,8 @@ pub async fn update_remote_profile(
     ).await?;
     let mut profile_data = ProfileUpdateData {
         username: actor.preferred_username.clone(),
-        hostname: webfinger_hostname,
+        hostname: server_hostname,
+        webfinger_hostname: webfinger_hostname,
         display_name: actor.name.clone(),
         bio: actor.summary.clone(),
         bio_source: actor.summary.clone(),