+++ mitra_models/src/profiles/queries.rs
@@ -157,10 +157,10 @@ pub async fn create_profile(
     profile_data.check_consistency()?;
     let transaction = db_client.transaction().await?;
     let profile_id = generate_ulid();
-    if let WebfingerHostname::Remote(ref hostname) = profile_data.hostname {
+    if let Some(ref hostname) = profile_data.hostname {
         create_instance(&transaction, hostname).await?;
     };
-    match profile_data.hostname {
+    match profile_data.webfinger_hostname {
         WebfingerHostname::Local => (),
         WebfingerHostname::Remote(ref hostname) => {
             prevent_acct_conflict(
@@ -199,8 +199,8 @@ pub async fn create_profile(
         &[
             &profile_id,
             &profile_data.username,
-            &profile_data.hostname.as_str(),
-            &profile_data.hostname.as_str(),
+            &profile_data.hostname,
+            &profile_data.webfinger_hostname.as_str(),
             &profile_data.display_name,
             &profile_data.bio,
             &profile_data.avatar,
@@ -253,17 +253,15 @@ pub async fn update_profile(
         .flatten()
         .filter_map(|image| image.clone().into_file_name())
         .collect();
-    if profile_data.hostname.as_str() != profile.hostname.as_deref() &&
-        !profile.is_portable()
-    {
-        // Only portable actors can change hostname
+    if profile_data.hostname != profile.hostname && !profile.is_portable() {
+        // Only portable actors can change server hostname
         return Err(DatabaseTypeError.into());
     };
-    if let WebfingerHostname::Remote(ref hostname) = profile_data.hostname {
+    if let Some(ref hostname) = profile_data.hostname {
         create_instance(&transaction, hostname).await?;
     };
 
-    match profile_data.hostname {
+    match profile_data.webfinger_hostname {
         WebfingerHostname::Local => (),
         WebfingerHostname::Remote(ref hostname) => {
             prevent_acct_conflict(
@@ -304,8 +302,8 @@ pub async fn update_profile(
         ",
         &[
             &profile_data.username,
-            &profile_data.hostname.as_str(),
-            &profile_data.hostname.as_str(),
+            &profile_data.hostname,
+            &profile_data.webfinger_hostname.as_str(),
             &profile_data.display_name,
             &profile_data.bio,
             &profile_data.bio_source,
@@ -1101,7 +1099,8 @@ mod tests {
     async fn test_create_profile_remote() {
         let profile_data = ProfileCreateData {
             username: "test".to_string(),
-            hostname: WebfingerHostname::Remote("example.com".to_string()),
+            hostname: Some("example.com".to_owned()),
+            webfinger_hostname: WebfingerHostname::Remote("example.com".to_string()),
             public_keys: vec![DbActorKey::default()],
             actor_json: Some(DbActor::for_test("https://example.com/users/test")),
             ..Default::default()
@@ -1120,6 +1119,28 @@ mod tests {
             profile.actor_id.unwrap(),
             "https://example.com/users/test",
         );
+    }
+
+    #[tokio::test]
+    #[serial]
+    async fn test_create_profile_remote_with_split_domain() {
+        let profile_data = ProfileCreateData {
+            username: "test".to_string(),
+            hostname: Some("ap.example.com".to_owned()),
+            webfinger_hostname: WebfingerHostname::Remote("example.com".to_string()),
+            public_keys: vec![DbActorKey::default()],
+            actor_json: Some(DbActor::for_test("https://example.com/users/test")),
+            ..Default::default()
+        };
+        let db_client = &mut create_test_database().await;
+        let profile = create_profile(db_client, profile_data).await.unwrap();
+        profile.check_consistency().unwrap();
+        assert_eq!(profile.hostname.as_ref().unwrap(), "ap.example.com");
+        assert_eq!(
+            profile.webfinger_hostname(),
+            WebfingerHostname::Remote("example.com".to_owned()),
+        );
+        assert_eq!(profile.acct.unwrap(), "test@example.com");
     }
 
     #[tokio::test]
@@ -1150,7 +1171,8 @@ mod tests {
         let actor_id = "https://example.com/users/test";
         let profile_data_1 = ProfileCreateData {
             username: "test-1".to_string(),
-            hostname: WebfingerHostname::Remote("example.com".to_string()),
+            hostname: Some("example.com".to_owned()),
+            webfinger_hostname: WebfingerHostname::Remote("example.com".to_string()),
             public_keys: vec![DbActorKey::default()],
             actor_json: Some(DbActor::for_test(actor_id)),
             ..Default::default()
@@ -1158,7 +1180,8 @@ mod tests {
         create_profile(db_client, profile_data_1).await.unwrap();
         let profile_data_2 = ProfileCreateData {
             username: "test-2".to_string(),
-            hostname: WebfingerHostname::Remote("example.com".to_string()),
+            hostname: Some("example.com".to_owned()),
+            webfinger_hostname: WebfingerHostname::Remote("example.com".to_string()),
             public_keys: vec![DbActorKey::default()],
             actor_json: Some(DbActor::for_test(actor_id)),
             ..Default::default()
@@ -1173,7 +1196,8 @@ mod tests {
         let db_client = &mut create_test_database().await;
         let profile_data_1 = ProfileCreateData {
             username: "test".to_string(),
-            hostname: WebfingerHostname::Remote("social.example".to_string()),
+            hostname: Some("social.example".to_owned()),
+            webfinger_hostname: WebfingerHostname::Remote("social.example".to_string()),
             public_keys: vec![DbActorKey::default()],
             actor_json: Some(DbActor::for_test("https://social.example/users/1")),
             ..Default::default()
@@ -1182,7 +1206,8 @@ mod tests {
         assert_eq!(profile_1.acct.unwrap(), "test@social.example");
         let profile_data_2 = ProfileCreateData {
             username: "test".to_string(),
-            hostname: WebfingerHostname::Remote("social.example".to_string()),
+            hostname: Some("social.example".to_owned()),
+            webfinger_hostname: WebfingerHostname::Remote("social.example".to_string()),
             public_keys: vec![DbActorKey::default()],
             actor_json: Some(DbActor::for_test("https://social.example/users/2")),
             ..Default::default()