+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `--user-agent` parameter to `fetch-object` command.
 - Add `quote` property to `Note` object when post contains a quote.
 - Convert `Image` objects into posts with an attached image.
+- Discover real webfinger address when importing an actor.
 
 ### Changed