+++ apx_sdk/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added documentation to `apx_sdk::addresses` module.
 - Implemented `Hash` for `WebfingerAddress`.
 - Added `actor_id` method to `JsonResourceDescriptor` type.
+- Added `is_subdomain_of` function to `hostname` module.
 
 ### Changed